Edits to Consider
There was a lot of inconsistence grammar and the space issue and how sometimes Kin wasn't capitalized. I would go and double check it some more. 

Added README suggestions to add. Most of them have to do with consistency. They are in brackets that say [Don't push....]. They are just suggestions.